src/requirements.txt
absl-py==0.9.0
appnope==0.1.0
asn1crypto==1.3.0
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.2
bert-embedding==1.0.1
bleach==3.1.0
blinker==1.4
blis==0.4.1
boto==2.49.0
boto3==1.12.6
botocore==1.15.6
cachetools==4.0.0
catalogue==1.0.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.8
cycler==0.10.0
cymem==2.0.3
decorator==4.4.2
defusedxml==0.6.0
dnspython==1.16.0
docutils==0.15.2
entrypoints==0.3
filelock==3.0.12
fuzzywuzzy==0.18.0
gast==0.2.2
gensim==3.8.2
gluonnlp==0.6.0
google-auth==1.13.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
graphviz==0.8.4
grpcio==1.27.2
h5py==2.10.0
idna==2.9
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
jmespath==0.9.5
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.2
jupyter-console==6.1.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.6.3
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
leveldb==0.201
lxml==4.5.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.1.3
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
murmurhash==1.0.2
mwparserfromhell==0.5.4
mxnet==1.4.0
mxnet-cu100==1.5.1.post0
mysql-connector-python==8.0.19
nb-conda-kernels==2.2.3
nbconvert==5.6.1
nbformat==5.0.4
nltk==3.4.5
notebook==6.0.3
numpy==1.18.3
oauthlib==3.1.0
opt-einsum==3.1.0
pandas==1.0.1
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
plac==1.1.3
preshed==3.0.2
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==3.11.4
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
python-Levenshtein==0.12.0
pytz==2019.3
PyYAML==5.3.1
pyzmq==18.1.1
qtconsole==4.7.1
QtPy==1.9.0
quickumls==1.3.0.post4
quickumls-simstring==1.1.5.post1
regex==2020.2.20
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
sacremoses==0.0.38
scikit-learn==0.22.1
scipy==1.4.1
semantic-version==2.8.4
Send2Trash==1.5.0
-e git+https://github.com/UKPLab/sentence-transformers@7fa3f0df28b3ae2a1d7353602de4e020e5f9e8d0#egg=sentence_transformers
sentencepiece==0.1.85
setuptools-rust==0.10.6
six==1.14.0
sklearn==0.0
smart-open==1.11.1
soupsieve==2.0
spacy==2.2.3
SQLAlchemy==1.3.13
srsly==1.0.1
tb-nightly==1.14.0a20190603
tensorboard==2.1.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
tensorflow-gpu==2.0.0b1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
textblob==0.15.3
tf-estimator-nightly==1.14.0.dev2019060501
thinc==7.3.1
tika==1.23.1
tokenizers==0.0.11
toml==0.10.0
torch==1.4.0
tornado==6.0.4
tqdm==4.43.0
traitlets==4.3.3
transformers==2.3.0
typing==3.6.6
Unidecode==1.1.1
urllib3==1.25.8
wasabi==0.6.0
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wrapt==1.12.1
yapf==0.29.0
zipp==2.2.0
